moved post-redux to separate file
