+++ src/redux/reducers/index.js
@@ -1,0 +1,9 @@
+import {combineReducers} from 'redux';
+
+import {postsReducer} from './postsReducer';
+import {playerReducer} from './playerReducer';
+
+export default combineReducers({
+  postsReducer,
+  playerReducer
+});
